Remove IRenderEngine and RenderWindow only use SDL Renderer API
