+++ src/Aquila.Demo/Examples/Graphics/HelloTriangle.cs
@@ -4,8 +4,8 @@
 using KappaDuck.Aquila;
 using KappaDuck.Aquila.Events;
 using KappaDuck.Aquila.Geometry;
-using KappaDuck.Aquila.Graphics;
 using KappaDuck.Aquila.Graphics.Primitives;
+using KappaDuck.Aquila.Graphics.Rendering;
 using KappaDuck.Aquila.System;
 using KappaDuck.Aquila.Video.Windows;
 using System.Drawing;
